requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
edge-tts==7.0.2
ffmpeg-python==0.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
future==1.0.0
fuzzywuzzy==0.18.0
idna==3.10
Jinja2==3.1.6
Levenshtein==0.27.1
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
openai-whisper==20240930
propcache==0.3.1
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
srt==3.5.3
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
